backend/requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.51.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
mypy-extensions==0.4.3
mysqlclient==2.1.1
oauthlib==3.2.1
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
